+++ CHANGELOG.md
@@ -3,5 +3,18 @@
  # Version 0.0.5
  - Insert Check Legacy Dir/Filter Command
   # Version 0.0.7
- - Fix Check Legacy Dir/Filter Command , Read Loop On Entity
- 
+ - Fix Check Legacy Dir/Filter Command , Read Loop On Entity 1 Level
+ - Fix Upper case first letter transalte
+ # Version 0.0.8
+ - Fix Check Legacy Dir/Filter Command Error
+ - Fix conver to excel Enttiy to left
+ # Version 0.0.9
+ - Add options CheckLagacy when Save
+ # Version 0.0.10
+ - Fix Excel
+ - Fix settings When install extension
+ - Auto install XML-RedHat and XML-Mike Brugh When installed extension
+ # Version 0.0.11
+ - Turn off some notify
+  # Version 0.0.12
+ - Fix check legacy